fix "absolu" bug

+++ src/Lesko/CodeAnalysis/Symbols/BuiltinFunctions.cs
@@ -12,7 +12,7 @@ namespace Lesko.CodeAnalysis.Symbols
         public static readonly FunctionSymbol Input = new FunctionSymbol("lire", ImmutableArray<ParameterSymbol>.Empty, TypeSymbol.String);
         public static readonly FunctionSymbol Rnd = new FunctionSymbol("alea", ImmutableArray.Create(new ParameterSymbol("max", TypeSymbol.Int)), TypeSymbol.Int);
 
-        public static readonly FunctionSymbol abs = new FunctionSymbol("absolu", ImmutableArray.Create(new ParameterSymbol("number", TypeSymbol.Int)), TypeSymbol.Int);
+        public static readonly FunctionSymbol abs = new FunctionSymbol("absolu", ImmutableArray.Create(new ParameterSymbol("number", TypeSymbol.Float)), TypeSymbol.Float);
         public static readonly FunctionSymbol sqrt = new FunctionSymbol("racine", ImmutableArray.Create(new ParameterSymbol("number", TypeSymbol.Float)), TypeSymbol.Float);
         public static readonly FunctionSymbol round = new FunctionSymbol("arrondir", ImmutableArray.Create(new ParameterSymbol("number", TypeSymbol.Float)), TypeSymbol.Float);
